requirements.txt
python-dotenv
colorama
requests
pycryptodome
pycrypto